refactor: move CSV export functions to CSVExporter
